+++ DimensionWars_Server/DimensionWars_Server/stdafx.h
@@ -394,8 +394,8 @@ struct SOCKETINFO
 	XMFLOAT3 m_Right;
 	XMFLOAT3 m_Up;
 	XMFLOAT3 m_Look;
-	float hp = MAX_HP;
-	float sp = MAX_SP;
+	unsigned short hp = MAX_HP;
+	unsigned short sp = MAX_SP;
 
 	// 투사체 정보
 	bool tile_life = false;
